Move DataUtil, MathUtils, PropertyUtils, CompressingIdGenerator to commonMain
- Replace java.util.Random/SecureRandom with kotlin.random.Random
- Replace Integer.toString(n, radix) with n.toString(radix)
- Replace java.lang.Double.valueOf() with direct call
- Add SizeBoundVector.contains() override to fix KMP IR compiler issue
  when ArrayList subclasses coexist with commonMain files in same package

4 utility files now in commonMain, accessible from both JVM and iOS.

+++ commcare-core/src/main/java/org/javarosa/xpath/expr/XPathRandomFunc.kt
@@ -17,7 +17,7 @@ open class XPathRandomFunc : XPathFuncExpr, VolatileXPathFuncExpr {
 
     override fun evalBody(model: DataInstance<*>?, evalContext: EvaluationContext, evaluatedArgs: Array<Any?>): Any {
         //calculated expressions may be recomputed w/o warning! use with caution!!
-        return java.lang.Double.valueOf(MathUtils.getRand().nextDouble())
+        return MathUtils.getRand().nextDouble()
     }
 
     override fun rootExpressionTypeIsCacheable(): Boolean {
